Separated Homepage and HomeSuggestion, changed Profile to ProfileHeader
